requirements.txt
elasticsearch
fastapi
langchain
langchain-openai
pgvector
polars
psycopg2-binary
pydantic
python-dotenv
qdrant-client
sqlalchemy
structlog
redis
